Create shim for YAML template declarations
Pandoc doesn't have a way to  specify which template file should be used
within the markdown front-matter, so I'm using `sed` to see if a
template key exists, and if it does, using that as the argument for the
`--template` option.

Hopefully this will get added in the future. Following Pandoc
Issue #5870 to see what happens.

https://github.com/jgm/pandoc/issues/5870