Automate creation of `nextProject` links
No more having to specify next project in the markdown front-matter

Expands on commit c9912b

A new Lua filter checks to see if the content file is part of the
projects list (if it has `position`). It then uses a generated YAML list
(`./project-list.yml`, created by `scripts/create-project-list.js`)
to find the next project relative to it's own position. If it's the last
project, the next project is the first one.