Allow subdirectories in image directory
Getting the build script to handle this was more trouble than it should
have been. `imagemin` has a 4 year old merge request[1] to handle
recursing into subdirectories...
As a work around when images are batched, the script loops over all the
files in the image dir, running them in parallel groups of 4 (to speed
things up, but not lock up the CPU).

1: https://github.com/imagemin/imagemin/pull/353